dmabuf: Avoid division by 0
When opening mks for an already running VM, the QEMU dbus backend has no
way to tell to the VM refresh the display and submit a new UpdateDMABuf.

So by the time size_allocate is called, the height and width of the paintable are
still set to 0.

This should fix the size allocate warning we see now until we figure out a proper
fix from the QEMU side.